requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
scikit-learn>=1.1.0

# NLP Libraries
nltk>=3.8
spacy>=3.4.0
transformers>=4.20.0
torch>=1.12.0
sentence-transformers>=2.2.0
textblob>=0.17.0
wordcloud>=1.9.0
datasets>=2.0.0

# Topic Modeling
bertopic>=0.12.0
gensim>=4.2.0
umap-learn>=0.5.0
hdbscan>=0.8.0

# Text Processing and Web Scraping
beautifulsoup4>=4.11.0
requests>=2.28.0
tqdm>=4.64.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.15.0
ipywidgets>=8.0.0

# Evaluation and Metrics
rouge-score>=0.1.0
bert-score>=0.3.0
networkx>=2.8.0

# Visualization
pyLDAvis>=3.3.0

 # SpaCy Model
 en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl)